requirements.txt
fastapi
uvicorn
jinja2
pydantic
langgraph
langchain
langchain-ollama
langchain-core
python-multipart
motor
pymongo
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dateutil
itsdangerous